+++ components/ChoiceAnswer/SingleChoiceAnswer.tsx
@@ -1,4 +1,4 @@
-import { useMemo, CSSProperties } from 'react';
+import { useMemo } from 'react';
 import Radio from '@mui/material/Radio';
 import RadioGroup from '@mui/material/RadioGroup';
 import Typography from '@mui/material/Typography';
@@ -6,17 +6,8 @@ import FormControl from '@mui/material/FormControl';
 import FormControlLabel from '@mui/material/FormControlLabel';
 import ThemeProvider from '@mui/material/styles/ThemeProvider';
 import { ViewTypeEnum } from '../../enums';
-import useCustomTheme from '../../useCustomTheme';
-
-export interface SingleChoiceAnswerProps {
-  name?: CSSProperties['color'];
-  color?: string;
-  viewType: ViewTypeEnum;
-  options: Array<{ label: string; value: string }>;
-  value: string;
-  required?: boolean;
-  onChange: (value: string) => void;
-}
+import useCustomTheme from '../../hooks/useCustomTheme';
+import type { SingleChoiceAnswerProps } from './types';
 
 function SingleChoiseAnswer(props: SingleChoiceAnswerProps) {
   const { options, color, value, name, viewType, required, onChange } = props;
