+++ components/ChoiceAnswer/index.tsx
@@ -1,6 +1,9 @@
 export { default as EditChoiceItems } from './EditChoiceItems';
-export type { EditChoiceItemsProps } from './EditChoiceItems';
 export { default as SingleChoiceAnswer } from './SingleChoiceAnswer';
-export type { SingleChoiceAnswerProps } from './SingleChoiceAnswer';
 export { default as MultiChoiceAnswer } from './MultiChoiceAnswer';
-export type { MultiChoiceAnswerProps } from './MultiChoiceAnswer';
+
+export type {
+  EditChoiceItemsProps,
+  SingleChoiceAnswerProps,
+  MultiChoiceAnswerProps
+} from './types';
