+++ components/QuestionGroup/types.ts
@@ -1,0 +1,17 @@
+import type { CSSProperties } from 'react';
+import type { ScoringProps } from '../Scoring';
+import type { EditChoiceItemsProps } from '../ChoiceAnswer';
+import type { AnswerTypeSelectorProps } from '../AnswerTypeSelector';
+
+export interface QuestionGroupProps {
+  question: string;
+  description?: string;
+  color?: CSSProperties['color'];
+  questionLabel?: string;
+  inputLabel?: string;
+  answerType: AnswerTypeSelectorProps;
+  choices?: EditChoiceItemsProps;
+  showScoring?: boolean;
+  scoring?: ScoringProps;
+  onChangeQuestion: (value: string) => void;
+}
