+++ components/TextAnswer/types.ts
@@ -1,0 +1,16 @@
+import type { CSSProperties } from 'react';
+import type { TextFieldProps } from '@mui/material/TextField';
+import type { AnswerTypeEnum, ViewTypeEnum } from '../../enums';
+
+type TextProps = Omit<Omit<TextFieldProps, 'onChange'>, 'color'>;
+
+export interface TextAnswerProps extends TextProps {
+  value: string;
+  color?: CSSProperties['color'];
+  maxLength?: number;
+  minLength?: number;
+  required?: boolean;
+  viewType: ViewTypeEnum;
+  answerType: AnswerTypeEnum;
+  onChange: (value: string) => void;
+}
