+++ components/index.tsx
@@ -1,13 +1,17 @@
 export * from './ChoiceAnswer';
-export { default as LinkAnswer } from './LinkAnswer';
-export type { LinkAnswerProps } from './LinkAnswer';
-export { default as TextAnswer } from './TextAnswer';
-export type { TextAnswerProps } from './TextAnswer';
-export { default as UploaderAnswer } from './UploaderAnswer';
-export type { UploaderAnswerProps } from './UploaderAnswer';
-export { default as AnswerTypeSelector } from './AnswerTypeSelector';
-export type { AnswerTypeSelectorProps } from './AnswerTypeSelector';
-export { default as Scoring } from './Scoring';
-export type { ScoringProps } from './Scoring';
+
 export { default as AnswerGroup } from './AnswerGroup';
+export { default as AnswerTypeSelector } from './AnswerTypeSelector';
+export { default as LinkAnswer } from './LinkAnswer';
+export { default as TextAnswer } from './TextAnswer';
+export { default as Scoring } from './Scoring';
 export { default as QuestionGroup } from './QuestionGroup';
+export { default as UploaderAnswer } from './UploaderAnswer';
+
+export type { AnswerGroupProps } from './AnswerGroup';
+export type { AnswerTypeSelectorProps } from './AnswerTypeSelector';
+export type { LinkAnswerProps } from './LinkAnswer';
+export type { TextAnswerProps } from './TextAnswer';
+export type { ScoringProps } from './Scoring';
+export type { QuestionGroupProps } from './QuestionGroup';
+export type { UploaderAnswerProps } from './UploaderAnswer';
