+++ index.tsx
@@ -1,7 +1,7 @@
 /**
  * =======================================
  *    Components For Custom Questions
- *    Created At Oct, 2021 by Janden
+ *    Created At June, 2024 by Janden
  * =======================================
  */
