+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@rinxun/custom-questions",
-  "version": "1.3.0-alpha.4",
+  "version": "1.3.0",
   "description": "It can help you to build some React custom question components for editing and answering",
   "keywords": [
     "react",
